requirements.txt
pydantic-ai[openai]>=0.0.36
openai>=1.40.0
streamlit>=1.40.0
duckdb>=1.1.0
faiss-cpu>=1.7.0
PyMuPDF>=1.24.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
plotly>=5.24.0
numpy>=1.26.0
tiktoken>=0.7.0
pytest>=8.0.0
